chore: migrate common code to acceptance runner
